requirements.txt
anyio==4.5.2
certifi==2025.1.31
dnspython==2.6.1
dotenv==0.9.9
exceptiongroup==1.2.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
pymongo==4.10.1
python-dotenv==1.0.1
python-telegram-bot==21.6
sniffio==1.3.1
typing-extensions==4.13.0
